int to byte

+++ src/DotRecast.Detour.TileCache/DtTileCacheBuilder.cs
@@ -31,8 +31,8 @@ namespace DotRecast.Detour.TileCache
 {
     public static class DtTileCacheBuilder
     {
-        public const int DT_TILECACHE_NULL_AREA = 0;
-        public const int DT_TILECACHE_WALKABLE_AREA = 63;
+        public const byte DT_TILECACHE_NULL_AREA = 0;
+        public const byte DT_TILECACHE_WALKABLE_AREA = 63;
         public const int DT_TILECACHE_NULL_IDX = 0xffff;
 
         private static readonly int[] DirOffsetX = { -1, 0, 1, 0, };
